Update the file organization
