+++ src/util/util_string.hpp
